refactor util files
