+++ apps/roam/src/utils/advancedSearchFooterUtils.ts
@@ -3,7 +3,12 @@ import getPageTitleByPageUid from "roamjs-components/queries/getPageTitleByPageU
 import getTextByBlockUid from "roamjs-components/queries/getTextByBlockUid";
 import updateBlock from "roamjs-components/writes/updateBlock";
 import { stripTypePrefix } from "~/components/AdvancedNodeSearchDialog/utils";
-import { getBlockSelection } from "~/utils/getBlockSelection";
+
+export type BlockSelection = {
+  selectionStart: number;
+  selectionEnd: number;
+  selectedText: string;
+};
 
 export type InsertTarget = {
   blockUid: string;
@@ -11,6 +16,44 @@ export type InsertTarget = {
 };
 
 const DEFAULT_WINDOW_ID = "main-window";
+
+export const getBlockSelection = (uid: string): BlockSelection => {
+  const activeElement = document.activeElement;
+  const isFocusedTextarea =
+    activeElement instanceof HTMLTextAreaElement &&
+    activeElement.classList.contains("rm-block-input") &&
+    getUids(activeElement).blockUid === uid;
+  if (isFocusedTextarea) {
+    return {
+      selectionStart: activeElement.selectionStart,
+      selectionEnd: activeElement.selectionEnd,
+      selectedText: activeElement.value.substring(
+        activeElement.selectionStart,
+        activeElement.selectionEnd,
+      ),
+    };
+  }
+  const textareas = document.querySelectorAll("textarea.rm-block-input");
+  for (const el of textareas) {
+    const textarea = el as HTMLTextAreaElement;
+    if (getUids(textarea).blockUid === uid) {
+      return {
+        selectionStart: textarea.selectionStart,
+        selectionEnd: textarea.selectionEnd,
+        selectedText: textarea.value.substring(
+          textarea.selectionStart,
+          textarea.selectionEnd,
+        ),
+      };
+    }
+  }
+  const textLength = (getTextByBlockUid(uid) || "").length;
+  return {
+    selectionStart: textLength,
+    selectionEnd: textLength,
+    selectedText: "",
+  };
+};
 
 const insertTargetFromFocusedBlock = (): InsertTarget | null => {
   const focusedBlock = window.roamAlphaAPI.ui.getFocusedBlock();
